requirements.txt
amqp==1.4.9
anyio==3.5.0
anyjson==0.3.3
appdirs==1.4.4
APScheduler==3.6.3
argh==0.26.2
asgiref==3.5.2
backports.csv==1.0.7
backports.zoneinfo==0.2.1
beautifulsoup4==4.8.1
billiard==3.3.0.23
boto==2.49.0
boto3==1.10.36
botocore==1.13.36
bs4==0.0.1
cachetools==4.2.2
celery==3.1.26.post2
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
clarifai==2.6.2
Click==7.0
configparser==5.0.1
confusable-homoglyphs==3.2.0
cryptography==39.0.0
cssselect==1.1.0
defusedxml==0.6.0
deta==1.1.0
diff-match-patch==20181111
distlib==0.3.4
dj-database-url==1.0.0
Django==4.1.2
django-admin-lightweight-date-hierarchy==1.0.1
django-admin-ordering==0.12.0
django-admin-shortcuts==2.0.0
django-anymail==8.6
django-appconf==1.0.5
django-autocomplete-light==3.9.4
django-celery==3.3.1
django-ckeditor==5.8.0
django-compressor==4.1
django-cors-headers==3.13.0
django-cryptography==1.1
django-csp==3.7
django-debug-toolbar==3.4.0
django-extensions==3.2.1
django-filter==21.1
django-google-analytics-app==4.4.0
django-heroku==0.3.1
django-htmx==1.13.0
django-import-export==2.8.0
django-js-asset==1.2.2
django-livereload-server==0.3.2
django-modeladmin-reorder==0.3.1
django-permissions-policy==4.15.0
django-registration==3.3
django-rest-framework==0.1.0
django-storages==1.8
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
docutils==0.15.2
EasyProcess==0.3
emoji==1.2.0
et-xmlfile==1.0.1
fake-useragent==0.1.11
fastapi==0.75.0
filelock==3.7.0
frappe-bench==5.2.1
future==0.18.2
gitdb2==2.0.6
GitPython==2.1.15
gmplot==1.4.1
googleapis-common-protos==1.52.0
grpcio==1.35.0
gunicorn==20.1.0
honcho==1.0.1
idna==2.8
importlib-metadata==4.10.1
imutils==0.5.4
instapy==0.6.13
jdcal==1.4.1
Jinja2==2.10.3
jmespath==0.9.4
jsonschema==2.6.0
kombu==3.0.37
lxml==4.7.1
MarkupPy==1.14
MarkupSafe==1.1.1
MeaningCloud-python==2.0.0
numpy==1.21.3
odfpy==1.4.0
opencv-python==4.5.4.58
openpyxl==3.0.2
pandas==1.3.4
parse==1.19.0
pathtools==0.1.2
Pillow==7.2.0
pipenv==2021.11.15
platformdirs==2.5.2
plyer==2.0.0
protobuf==3.14.0
psycopg2==2.9.1
pycparser==2.21
pydantic==1.9.0
pyee==8.2.2
pyppeteer==1.0.2
pyquery==1.4.3
pytesseract==0.3.8
python-crontab==2.4.0
python-dateutil==2.8.0
python-http-client==3.2.1
python-telegram-bot==13.10
pytz==2019.3
PyVirtualDisplay==2.0
PyYAML==5.2
qrcode==6.1
rcssmin==1.1.0
regex==2020.11.13
requests==2.22.0
requests-html==0.10.0
rjsmin==1.2.0
s3transfer==0.2.1
selenium==3.141.0
semantic-version==2.8.2
sendgrid==3.6.5
sendgrid-django==4.2.0
six==1.16.0
smmap==3.0.4
smmap2==3.0.1
sniffio==1.2.0
soupsieve==1.9.5
speedtest-cli==2.1.3
sqlparse==0.4.3
starlette==0.17.1
structlog==19.2.0
tablib==3.2.0
tornado==6.1
tqdm==4.56.0
typing-extensions==4.1.1
tzlocal==2.1
urllib3==1.25.7
virtualenv==20.14.1
virtualenv-clone==0.5.6
w3lib==1.22.0
watchdog==0.9.0
webdriverdownloader==1.1.0.3
websocket-client==0.56.0
websockets==10.1
whitenoise==6.2.0
xlrd==1.2.0
xlwt==1.3.0
youtube-dl==2021.12.17
zipp==3.7.0
